requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Flask==0.12.2
flask-heroku==0.1.9
Flask-SQLAlchemy==2.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
psycopg2==2.7.3.1
SQLAlchemy==1.1.14
Werkzeug==0.12.2
gunicorn==19.7.1
python-dotenv==0.7.1
oauthlib==2.0.4
requests-oauthlib==0.8.0
urllib3==1.22
